(BKR-623) Rename runner/beaker to runner/native, ...
... and bring `Beaker::TestCase` into the fold.

Prior to this, `Beaker::TestCase` was still global. In moving it to a nested
`Beaker::Runner::Beaker::TestCase` I found that a few tests were starting to
fail. Further investigation showed that references to `Beaker::*` classes
inside tests were being resolved as `Beaker::Runner::Beaker::*`, and throwing
constant missing exceptions. Moving `Beaker::Runner::Beaker` to
`Beaker::Runner::Native` allows ruby's nested class resolution algorithm
to find `::Beaker` instead of a nested `Beaker` module constant.
